requirements.txt
annotated-types==0.7.0
mypy_extensions==1.1.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pandera==0.23.1
phonenumbers==9.0.4
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typeguard==4.4.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
